Solidity linting

+++ contracts/CommuniftyNFT.sol
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.8.0;
 
-import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
-import "@openzeppelin/contracts/utils/math/SafeMath.sol";
-import "@openzeppelin/contracts/utils/Strings.sol";
-import "./interfaces/ICommuniftyNFT.sol";
+import '@openzeppelin/contracts/token/ERC721/ERC721.sol';
+import '@openzeppelin/contracts/utils/math/SafeMath.sol';
+import '@openzeppelin/contracts/utils/Strings.sol';
+import './interfaces/ICommuniftyNFT.sol';
 
 contract OwnableDelegateProxy {}
 
@@ -23,7 +23,8 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
     uint32 private constant MAX_OWNER_FEE = 950; // 95.0%
 
     uint8 public constant platformFee = 50; // 5.0%
-    address payable public constant platform = payable(0xAD10D4F9937D743cbEb1383B1D3A3AD15Ace75D6);
+    address payable public constant platform =
+        payable(0xAD10D4F9937D743cbEb1383B1D3A3AD15Ace75D6);
 
     uint32 public m = SLOPE_DENOM;
     uint32 public n = 1;
@@ -42,11 +43,11 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
     address public proxyRegistryAddress;
 
     uint256 private currentTokenId = 0;
-    string private baseURI = "";
+    string private baseURI = '';
 
     /**
-    * @dev Create the NFT contract from name symbol and baseURI
-    */
+     * @dev Create the NFT contract from name symbol and baseURI
+     */
     constructor(
         string memory _name,
         string memory _symbol,
@@ -86,7 +87,7 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
         emit PreMint(_count, _to);
     }
 
-    function nftReservePoolLength() external view returns (uint) {
+    function nftReservePoolLength() external view returns (uint256) {
         return nftReservePool.length;
     }
 
@@ -106,27 +107,38 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
     }
 
     function _setBaseURI(string memory baseURI_) private {
-      baseURI = baseURI_;
+        baseURI = baseURI_;
     }
 
-    function tokenURI(uint256 tokenId) public view override returns (string memory) {
-      require(_exists(tokenId), "ERC721Metadata: URI query for nonexistent token");
+    function tokenURI(uint256 tokenId)
+        public
+        view
+        override
+        returns (string memory)
+    {
+        require(
+            _exists(tokenId),
+            'ERC721Metadata: URI query for nonexistent token'
+        );
 
-      if (bytes(tokenURIs[tokenId]).length > 0) {
-        return tokenURIs[tokenId];
-      }
+        if (bytes(tokenURIs[tokenId]).length > 0) {
+            return tokenURIs[tokenId];
+        }
 
-      return bytes(baseURI).length > 0 ? string(abi.encodePacked(baseURI, tokenId.toString())) : "";
+        return
+            bytes(baseURI).length > 0
+                ? string(abi.encodePacked(baseURI, tokenId.toString()))
+                : '';
     }
 
     function setTokenURI(uint256 tokenId, string memory uri) external {
-      require(msg.sender == pairOwner, 'FORBIDDEN_PAIR_OWNER');
-      require(bytes(tokenURIs[tokenId]).length == 0, "URI_ALREADY_SET");
-      tokenURIs[tokenId] = uri;
+        require(msg.sender == pairOwner, 'FORBIDDEN_PAIR_OWNER');
+        require(bytes(tokenURIs[tokenId]).length == 0, 'URI_ALREADY_SET');
+        tokenURIs[tokenId] = uri;
     }
 
     function totalSupply() external view returns (uint256) {
-      return currentTokenId;
+        return currentTokenId;
     }
 
     /**
@@ -148,8 +160,8 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
     }
 
     /**
-    * @dev Transfer ownership of the pair's reserves and owner fees
-    */
+     * @dev Transfer ownership of the pair's reserves and owner fees
+     */
     function setPairOwner(address payable _nextOwner) external override {
         require(msg.sender == pairOwner, 'FORBIDDEN_PAIR_OWNER');
         require(_nextOwner != address(0), 'INVALID_PAIR_OWNER');
@@ -165,19 +177,22 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
     }
 
     /**
-    * @dev Function to close the market and retrun the ethReserve to the pair owner
-    */
+     * @dev Function to close the market and retrun the ethReserve to the pair owner
+     */
     function close() external override {
-        require(closeDeadline != 0 && block.timestamp >= closeDeadline, 'INVALID_DEADLINE');
+        require(
+            closeDeadline != 0 && block.timestamp >= closeDeadline,
+            'INVALID_DEADLINE'
+        );
         pairOwner.transfer(ethReserve);
         emit Close(msg.sender, ethReserve);
         ethReserve = 0;
     }
 
     /**
-    * @dev Function to remove fees attributed to owner.
-    * Fees for the owner are reset to 0 once called.
-    */
+     * @dev Function to remove fees attributed to owner.
+     * Fees for the owner are reset to 0 once called.
+     */
     function withdrawOwnerFees() external override {
         pairOwner.transfer(ownerFees);
         emit WithdrawOwnerFees(msg.sender, ownerFees);
@@ -185,23 +200,34 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
     }
 
     /**
-    * @dev Function to remove fees attributed to platform.
-    * Fees for the platform are reset to 0 once called.
-    */
+     * @dev Function to remove fees attributed to platform.
+     * Fees for the platform are reset to 0 once called.
+     */
     function withdrawPlatformFees() external override {
         platform.transfer(platformFees);
         emit WithdrawPlatformFees(msg.sender, platformFees);
         platformFees = 0;
     }
 
-    function buy(address payable _to) external payable override returns (uint256 tokenId) {
-        require(closeDeadline == 0 || block.timestamp < closeDeadline, 'MARKET_CLOSED');
-        uint basePrice = priceAtX(x);
+    function buy(address payable _to)
+        external
+        payable
+        override
+        returns (uint256 tokenId)
+    {
+        require(
+            closeDeadline == 0 || block.timestamp < closeDeadline,
+            'MARKET_CLOSED'
+        );
+        uint256 basePrice = priceAtX(x);
         // Determine if we can sell a token id from the reserve pool
         if (nftReservePoolIndex < nftReservePool.length) {
             tokenId = nftReservePool[nftReservePoolIndex];
-            require(_isApprovedOrOwner(address(this), tokenId), 'UNAPPROVED_BUY');
-             _transfer(address(this), _to, tokenId);
+            require(
+                _isApprovedOrOwner(address(this), tokenId),
+                'UNAPPROVED_BUY'
+            );
+            _transfer(address(this), _to, tokenId);
             nftReservePoolIndex++;
         } else {
             // Mint a new token to the recipient
@@ -211,8 +237,8 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
             _incrementTokenId();
         }
         // calculate fees,
-        uint platformShare = basePrice.mul(platformFee).div(1000);
-        uint ownerShare = basePrice.mul(ownerFee).div(1000);
+        uint256 platformShare = basePrice.mul(platformFee).div(1000);
+        uint256 ownerShare = basePrice.mul(ownerFee).div(1000);
         platformFees = platformFees.add(platformShare);
         ownerFees = ownerFees.add(ownerShare);
         // Update ETH reserve
@@ -220,7 +246,7 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
         // Update X
         x = x + 1;
         // check value vs purchase price, and refund if necessary
-        uint purchasePrice = basePrice.add(platformShare).add(ownerShare);
+        uint256 purchasePrice = basePrice.add(platformShare).add(ownerShare);
         require(msg.value >= purchasePrice, 'INSUFFICIENT_FUNDS');
         // Refund excess ETH
         if (msg.value > purchasePrice) {
@@ -230,21 +256,24 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
     }
 
     function sell(uint256 _tokenId, address payable _to) external override {
-        require(closeDeadline == 0 || block.timestamp < closeDeadline, 'MARKET_CLOSED');
+        require(
+            closeDeadline == 0 || block.timestamp < closeDeadline,
+            'MARKET_CLOSED'
+        );
         require(x > 1, 'INVALID_X');
-        uint salesPrice = priceAtX(x - 1);
+        uint256 salesPrice = priceAtX(x - 1);
         require(ethReserve >= salesPrice, 'INSUFFICIENT_RESERVE');
         // Send the token to this contract and add to reserve pool
         require(_isApprovedOrOwner(msg.sender, _tokenId), 'UNAPPROVED_SELL');
         _transfer(msg.sender, address(this), _tokenId);
         nftReservePool.push(_tokenId);
         // Calculate fees
-        uint platformShare = salesPrice.mul(platformFee).div(1000);
-        uint ownerShare = salesPrice.mul(ownerFee).div(1000);
+        uint256 platformShare = salesPrice.mul(platformFee).div(1000);
+        uint256 ownerShare = salesPrice.mul(ownerFee).div(1000);
         platformFees = platformFees.add(platformShare);
         ownerFees = ownerFees.add(ownerShare);
         // deduct fees from sales proceeds
-        uint saleProceeds = salesPrice.sub(platformShare).sub(ownerShare);
+        uint256 saleProceeds = salesPrice.sub(platformShare).sub(ownerShare);
         // Transfer sale proceeds
         _to.transfer(saleProceeds);
         // Update ETH reserve
@@ -255,26 +284,26 @@ contract CommuniftyNFT is ICommuniftyNFT, ERC721 {
     }
 
     function priceAtX(uint256 _x) public view returns (uint256) {
-        return m.mul( (_x ** n).mul(10 ** 18) ).div(SLOPE_DENOM);
+        return m.mul((_x**n).mul(10**18)).div(SLOPE_DENOM);
     }
 
     /**
-    * @dev Returns the current buy price of a single NFT, including fees
-    */
+     * @dev Returns the current buy price of a single NFT, including fees
+     */
     function buyPrice() public view override returns (uint256) {
         uint256 basePrice = priceAtX(x);
-        uint platformShare = basePrice.mul(platformFee).div(1000);
-        uint ownerShare = basePrice.mul(ownerFee).div(1000);
+        uint256 platformShare = basePrice.mul(platformFee).div(1000);
+        uint256 ownerShare = basePrice.mul(ownerFee).div(1000);
         return basePrice.add(platformShare).add(ownerShare);
     }
 
     /**
-    * @dev Returns the current sell proceeds of a single NFT, including fees
-    */
+     * @dev Returns the current sell proceeds of a single NFT, including fees
+     */
     function sellPrice() public view override returns (uint256) {
         uint256 basePrice = priceAtX(x - 1);
-        uint platformShare = basePrice.mul(platformFee).div(1000);
-        uint ownerShare = basePrice.mul(ownerFee).div(1000);
+        uint256 platformShare = basePrice.mul(platformFee).div(1000);
+        uint256 ownerShare = basePrice.mul(ownerFee).div(1000);
         return basePrice.sub(platformShare).sub(ownerShare);
     }
 }
